Move turn factory tests
